+++ src/tuya_cloud_service/netmgr/src/conn/netconn_wired.c
@@ -152,7 +152,7 @@ OPERATE_RET netconn_wired_set(netmgr_conn_config_type_e cmd, void *param)
 
     switch (cmd) {
     case NETCONN_CMD_PRI:
-        netmgr_wired->base.pri = *(int *)param;
+        netmgr_wired->base.pri = *(uint8_t *)param;
         /* Guarded like every other call to it in this file. Unreachable through
          * netmgr_conn_set(), which refuses before netmgr_init() has installed the
          * shim - but netconn_wired_set() is on the global public include path, so
@@ -189,7 +189,7 @@ OPERATE_RET netconn_wired_get(netmgr_conn_config_type_e cmd, void *param)
 
     switch (cmd) {
     case NETCONN_CMD_PRI:
-        *(int *)param = netmgr_wired->base.pri;
+        *(uint8_t *)param = netmgr_wired->base.pri;
         break;
     case NETCONN_CMD_IP:
         TUYA_CALL_ERR_RETURN(tal_wired_get_ip((NW_IP_S *)param));
